Reorganize existing content (files/projects)
This repository re-organization separates the concepts of the example
package and individual template files. The basic structure is to have
two root directories:

- file_templates: for single files and partial file (snippets)
- project_templates: for multi-file project templates (like the structure of a
  Git repository.

Each file/project templates has its own README, and can have its own
example file in addition to the template itself. The template is a
Jinja2 template that can be instantiated with cookiecutter.