Add example_subpackage stack_package example
This example is in addition to the regular `example/`. This new example
has a Python namespace of `lsst.example.subpackage`. Therefore it shows
off having an intermediate Python directory, with an __init__.py that
extends the namespace.

To make this possible, this example uses

- The Cookiecutter post gen hook to copy in another __init__.py file
- The new Scons Cookiecutter builder that lets us pass cookiecutter
  variable overrides from the SConstruct.

Since there are now two examples, it's necessary to explicitly name the
target directories. To ensure that the example is always rebuilt, we
wrap the target in a AlwaysBuild() command.