Add template for a root header file
This header file is typically used to simply include other header files.

The root header file's path and name are based on the package's
namespace. I couldn't put the template logic directly in the paths to
place the file with only Jinja/Cookiecutter. Instead, I resorted to
adding a post-gen hook that moves include/root.h to the appropriate
path.

Includes user docs