Add templatekit make subcommand
This make subcommand lets a user actually create a new file/snippet or
project from a template.

Either kind of template can be made from this command because I think
that will be simpler for the user.

For file templates, the output can either go to stdout or to a file.

For project templates, we use the standard cookiecutter workflow and
create the project either in the current working directory or in a
user-specified base directory.