Redesign module and pacakge homepages
Since packages that provide Python modules not longer have package
homepages, the module homepage now needs to take on responsibility for
linking to the underlying Git repository and Jira component.

I'm adding a new Contributing section to the module homepage that links
to the repository and Jira component. The section can also be expanded
with a toctree linking to additional topics on developing the module (as
opposed to _using_ the module).

The package homepage also now has a matching Contributing section.

In addition, I've formalized the Using section, and included it in the
template, though commented-out.