Switch to git diff --exit-code flag
The --exit-code flag is necessary so that the git diff command
returns a non-zero status code if scons-ing the examples produced
changes.

However git diff-files was produces "false positives" on Travis
(possibly because of changed file ownership? or modification date
metadata? not sure). However, git diff seems to just check if the
content has changed, which is exactly what we want.