Eliminate intermediate __init__.py
In Python 3, __init__.py files are no longer needed to establish a
package. This also means that the pkgutil.extent_path method is no
longer needed. Nor is the lsstimport package from the "base" EUPS
package.

This keeps the __init__.py of the module itself; this file is still
commonly used for establishing public namespaces so it's good to keep it
in the template.

Also remove the base dependency, since lsstimport is no longer used.