Add cookiecutter.base_package variable
This lets a user select the base package. LSST Science Pipelines
packages should use "base" as the base package, and that also implicitly
provides common third-party packages like Python, pybind11, doxygen, and
sconsUtils, along with the lsstimport infrastructure.

This seems to be the effective consensus of RFC-52.

Stacks that don't need all the features of base can opt for
`sconsUtils`, which is the minimum needed to build packages.