Use True/False strings instead of booleans
Cookiecutter wants True/False booleans to really be "True" or "False"
strings. It converts booleans automatically to strings in the Jinja
context for example. So to to avoid confusion, lets just use "True" and
"False" strings from the get-to in the cookiecutter.json file.

This prevents an issue where "False" was passed to a variable like
has_tasks, and since "False" isn't part of the set, it defaulted to the
True boolean.