Move troubleshooting out of getting started sec
The troubleshooting info was too prominent in its original location
inside the Getting Started docs; it pushed the list of templates too far
down the page.

Few users will need to read these docs (most set up their shell) and
click itself provides an error message itself that links to the
resolution. So I think moving the troubleshooting docs to the bottom of
the README is appropriate.