Add Script reference section to module homepage
This is the place that will list scripts for each module and like to
those script topic pages.

I've chosen not to make this a toggleable feature, and instead just have
it commented out by default. Making it a template variable would add too
many entires to the "Flavor" menu.