Add uses_dds variable to configure T&S tests
The "uses_dds" cookiecutter variable configures whether to add an extra
block to the tests/SConscript file that adds the OSPL_URI,
OPENSPLICE_LOC, and ADLINK_LICENSE variables to the SCons environment.
These are necessary for T&S packages that work with DDS (the base
technology used by SAL).

The "example_dds" examples demonstrates this configuration.

For Slack users, this configuration is available from the "DDS Package"
presets.