Add the basic structure of a safir-based app
Note: this doesn't include safir yet as a dependency because it isn't
pip-installable, and thus can't be hashed. We'll fix this up once an
initial version of safir makes it to PyPI.

Features:

- App has two parts: the root application that hosts
internally-accessible, and a subapp that hosts endpoints at a subpath.
This configuration lets us dynamically set the subpath based on the
application configuration.

- Starter for app configuration, as a dataclass, that accepts defaults
based on environment variables. Users can change this up to use a
configuration instead, or in addition; this is a good starting point.

- Set up the logging middleware.

- Starting point for application metadata that primarily gathers
metadata from the package (setup.cfg). This is a hopefully useful way of
single-sourcing this metadata.

The root routes expose this metadata.

Users can extend this metadata with their own fields.