Add tox-based testing set up
This tox+pre-commit set up is heavily inspired by

- https://hynek.me/articles/python-app-deps-2018/
- https://hynek.me/articles/testing-packaging/
- https://medium.com/staqu-dev-logs/keeping-python-code-clean-with-pre-commit-hooks-black-flake8-and-isort-cac8b01e0ea1

The main tox command runs, by default, pytest with coverage, coverage
reporting, a "typing" environemnt for mypy and a "linting" environment
that runs a set of pre-commit scripts (black, isort, and flake8).

What's neat about this set up is that these tests are done in isolated
environments, so you don't have to set up a virtual environment of your
own if you don't want to. The same test set up will also be the basis of
the CI script. And if you want, straight-up "pytest" still works.

Also, combining the code style testing with the pre-commit step, we
ensure that code style is both tested but automatically corrected
whenever code is committed. This makes maintaining consistent styling
fairly effortless.

End users of this template can of course remove certain aspects of the
testing/linting set up, such as mypy and black if they want.