Add GitHub Actions CI workflow
This features two jobs:

1. Test and lint the app by running tox

2. Build the docker image.

Both jobs feature caching. The tox environment is cached, and keyed to
the hash of the files that define dependencies.

The intermediate docker image is pushed to docker hub and pulled for new
builds. This lets us reuse a fully cached dependencies image for most
builds.