Enable show_error_codes for mypy
This provides the specific error code that can be used in a
type: ignore comment to suppress only that error.  That isn't
useful yet because flake8 will complain about those comments, but
that will be fixed in the next flake8 release and there's no harm
in ensuring this configuration setting is present in newly-created
bots.