Remove license key from setup.cfg
packaging.python.org says that standard licenses should be
indicated primarily via trove classifiers and the license field is
optional and intended for custom non-standard licenses.  Omit it
and rely on the trove classifiers, which avoids the question of what
the valid values for this field are.

Reference: https://packaging.python.org/guides/distributing-packages-using-setuptools/#license