[DM-23913] Quoting and black reformatting
Two small things after running the safir bootstrap guide for my
new project, since I wanted to give it a shot!

It looks like the project name needs to be quoted as a string,
since the modulename isn't really imported or used there.  Quoting
seems to work.

After running black the first time, for some reason it really wants
there to be an extra comma after the keyword on its own line.  So I
figured it might be good to fix that just so there's not a lot of
reformatting going on in the first commit.