Update roundtable_aiohttp_bot template
Add a new script that installs packages required only in the
dependencies image and install build-essential by default, since
this is often required for building Python extensions, particularly
when upgrading to a new Python version.

Use a test matrix of Python 3.8 and 3.9 by default.  Stop pinning
the version in tox and use whatever the default Python version is.
Remove or update a few more references to Python 3.7.

Update pre-commit hook versions to the latest.

Use the @v2 syntax for the GitHub Actions configuration, since all
the actions move the major version tag with each release.  This
avoids requiring updates for each minor release.