Update CI and Docker builds
Use the official pre-commit action instead of running pre-commit
with tox, since it automatically handles caching.  Add the Python
version to the tox cache key and remove .pre-commit-config.yaml.

Change the Docker build to only trigger on ticket branches and
tags to cut down on the number of spurious images we push.  Use the
official Docker actions and cache layers in GitHub rather than on
Docker Hub.

Use an external script to determine the tag of the Docker image.