Use pytest-cov for FastAPI coverage report
Rather than running pytest under coverage, use pytest-cov.  This
will make it easier for a project to switch to using xdist down the
road, and it avoids the combine step or leaving lots of coverage
files behind.

This does not use pytest-xdist since most small packages won't need
it and it makes the output less pretty.