Fix version number for FastAPI applications
The version number was incorrectly set to a development working
version instead of the tagged version when building from a tag.
Fix this by including tags in the checkout so that setuptools_scm
can use them, not excluding any files from the Docker copy, and
using a separate working directory from the /app directory that's
used by the FastAPI image.