Update how workflows run
Our previous approach of running CI workflows on push but not pull
request is fairly hostile to external contributors.  Enable
workflows for pull requests in the relevant templates.  We also
want to re-run CI workflows on tags and merges to master to confirm
nothing broke, so also enable push CI but exclude the branches we
use to create local pull requests to avoid running the CI workflow
twice and wasting some resources.

It's not a big deal if we run workflows twice occasionally, so this
list of exclusions doesn't need to be comprehensive.