Fix Docker builds on ticket branches
The change to run CI workflows on pull requests and to avoid running
them on pushes to branches that will be pull requests broke running
the Docker build because github.ref is different for pull request
workflows.  Fix this by looking at the head_ref instead, to trigger
Docker builds for pull requests from ticket branches.

It's not entirely clear form the GitHub Actions documentation
whether having a branches_ignore key prevents running the action
on tags, so be explicit.