Fix Docker tags for app workflows
In the roundtable_aiohttp_bot and fastapi_safir_app project
templates, fix the script to determine the Docker tag of a build
to work correctly with pull requests as well as pushes.  Add
Python 3.10 to the default test matrix for both app types.

In the roundtable_aiohttp_bot project template, set fetch-depth
to 0 when building Docker images (matching fastapi_safir_app) so
that setuptools_scm will generate correct version numbers.