Stop using FastAPI Docker image
For the fastapi_safir_app template, stop using the FastAPI Docker
image as a base since it's been deprecated upstream.  It uses
gunicorn to start a bunch of worker processes, which is unnecessary
in Kubernetes (one can just spawn more pods) and often not desired
for applications that use in-memory data structures and should only
run one copy.

Instead, base the Docker image on the normal Python base image and
run the application directly using uvicorn.  Remove gunicorn from
the default dependencies.