Stop using subapps for FastAPI apps
Unfortunately, FastAPI or Starlette has a bug where exceptions
thrown by subapps are swallowed and not logged by the parent app.
Since we can simulate what we were doing with subapps using
routers, use a single FastAPI app by default to avoid this problem.

Mark the internal top-level route as undocumented so that it won't
be added to the public-facing API documentation, matching the
previous behavior where the subapp was exposing its own separate
API documentation.

We do lose the internal API documentation for non-visible routes,
but this doesn't seem like a significant loss.