Stop using if TYPE_CHECKING in FastAPI template
Using this construct is optional in our code, but we're starting to
lean away from using it routinely since it can mask some errors.
The uses in the test suite of the FastAPI app are not significant
or necessary, so remove this construction there to not create a
confusing example.