Unpin documenteer in tech note template
Versions of documenteer prior to 0.6.0 no longer work because the
current version of Jinja2 is incompatible with the LaTeX extension
forced by the older version of Sphinx.  Documentation builds fail
with the error:

> Could not import extension sphinx.builders.latex (exception: cannot import name 'contextfunction' from 'jinja2'

Rather than move the pin, unpin documenteer in the hope that using
the latest release will be more likely to succeed than older
partly-pinned versions.