Configure Uvicorn logging in FastAPI app
Add a call to the new Safir configure_uvicorn_logging function in
the template for the FastAPI Safir app, and set a dependency bound
on Safir to ensure that function is available.  This will ensure
that Uvicorn logs for FastAPI apps are routed through structlog and
the context expected by Google's Cloud Logging is added.