Manually set metadata for non lsstdoc documents
The --lsstdoc option in lander v1 wasn't intended to be used for other
LaTeX document classes, and although it works sometimes, it can fail.
Let's try setting core metadata, the title and handle, manually. We'll
return to this with Lander v2 where we can either set metadata in a
file, or introduce specific tex parsers for SPIE, ADASS, and AASTeX
documents.