Drop command-like task topic support
This drops the "script_name" field from the task_topic file template,
and therefore remove command-line task support. The task topic can
continue to be used for Task and PipelineTask topics. In the future
we'll need to add improved support for PipelineTask and pipelines, but
this at least provides immediate removal of support for deprecated
CmdLineTask task docs.