Stop adding middleware in the startup hook
This pattern was originally copied from Gafaelfawr, where I had
done this to defer adding the XForwardedMiddleware until after
configuration parsing so that the list of proxies could be configured.
It turned out to be a bad idea, since during the test suite, where
the application is repeatedly started and stopped, a new copy of the
middleware was added each time, eventually making the application
extremely slow.

I fixed this in Gafaelfawr (by switching to a callable to construct
the app) and forgot to ever fix the template. Now, Starlette prohibits
doing this (for good reasons), so it no longer works with the current
versions of FastAPI and Starlette.

Move the middleware construction to the top level, which will be
fine for any application that doesn't need to customize it. Apps that
do need to customize the middleware will need to use the callable
pattern to construct the app.