Add path_prefix support to FastAPI template
Several of our applications have wanted to be able to override the
base URL of the application (to test a parallel installation, for
example), and while this was supported via overriding the package
name, that approach is limited. (For instance, it's awkward to use
that to configure a multipart path prefix such as /api/example.)

Follow times-square and jupyterlab-controller and add a path_prefix
configuration option and environment variable to explicitly set the
prefix.

Fix a few places where the potentially-uppercase package name was
used in command-line entry points or URLs instead of the lowercased
version.