Substitute name and GitHub org in more places
In the FastAPI Safir app and SQuaRE PyPI package templates, there
were a few places where the application or library name or the GitHub
organization weren't substituted from the template variables. Add
substitutions.

(The issue with organization was noticed via build issues with
a package based on the FastAPI Safir template in the lsst-dm
organization. That use case isn't supported, but this seemed worth
fixing anyway in case we have more organizations in the future.)