Add neophile support to project templates
Add support for neophile to both the fastapi_safir_app and
square_pypi_package project templates.

For FastAPI apps with frozen dependencies, this means running
neophile as a separate CI check on each pull request, running a
periodic CI check that runs neophile and then the test suite, and
adding a periodic GitHub Actions workflow to create a PR to update
pre-commit dependencies.

For SQuaRE PyPI packages, this means running only the periodic
workflow to create a PR to update pre-commit dependencies, since
the other two steps are related to frozen dependencies.

If the created package is in lsst-sqre, report failures of the
periodic GitHub Actions workflows to Slack.