Pin documenteer, add dependabot configuration
Pin documenteer to <1 for square_pypi_project and technote_rst so
that we have more intentional control over when we move to newer
versions of the documentation stack. Add dependabot configuration
for PyPI dependencies to technote_rst so that dependabot will create
PRs to move to newer versions of documenteer.