Add Ruff configuration to project templates
We're not yet ready to switch to Ruff by default for linting, but
add our current Ruff configuration to the templates to make it easy
to do so for those who want to. Sort the pyproject.toml tool.*
configurations to put build system stuff first and then all testing
and linting tools in alphabetical order.