Better error reporting from PyPI linkcheck
Sphinx linkcheck doesn't produce a summary of failed links, but it
does write one to output.txt. Move the linkcheck command into the
Makefile so that we can use shell trickery to display the contents
of that file on failure.

Add a help target as the default target in Makefile.