Shorten the CHANGELOG.md preambles
The previous version was unnecessarily verbose and included
developer information about using scriv that is not likely to be
of interest to the target audience. Omit the note about updating
dependencies for PyPI apps; those changes aren't done automatically
with PyPI apps and may be significant enough to note in the change
log.