Update author metadata in technotes
The rst and md technotes now use the author ID to correspond with the
authordb.yaml file in github.com/lsst/lsst-texmf. This allows us to add
more structured metadata about authors that is necessary for
crosswalking to bibliographic schemas like Zenodo or CITATION.cff.

Also add the organization field to technote.toml. And record
date_created as a native TOML date-time, rather than a string.