Omit orcid from technote.toml if empty
Previously if an author didn't have an orcid in the authordb.yaml file
(or didn't provide one when working manually with templatekit), the
empty string would be added as the value for the orcid in technote.toml.
This breaks technote builds because the Pydantic model won't validate an
empty string as an orcid.

In jinja syntax we can omit the orcid line without creating whitespace
issues by putting an if clause on the same line and using the "-"
operator to consume the empty line.

Also fix the "tuple" in address example value. There shouldn't be an
extra set of parentheses on the example address values.