Update the FastAPI template for current practices
Update for current FastAPI recommendations and warnings:

- Use a lifespan function instead of hooks
- Update to Pydantic v2
- Use Annotated for injected dependencies in route functions

Use Pydantic's support for environment variable prefixes to use
the same prefix for all environment variables used for
configuration, instead of using SAFIR_ for the default options.

Update the Makefile for current best practices:

- Add a make help default target
- Upgrade pip before updating dependencies
- Run pre-commit autoupdate during make update-deps
- Add --allow-unsafe to pip-compile because setuptools is normally
  needed (and we've not had any problems with it in the past)

Remove the dependency update CI job. We've found this is more
annoying than helpful; it's easier to update using make update at
the start of a development cycle.

Switch to Ruff for both linting and reformatting; remove the Black,
isort, and flake8 configurations; and update the Ruff configuration
with more exclusions and for 0.2.0.

Update the default and minimum version of Python to 3.12.