Update versions in FastAPI template
In the Dockerfile for the FastAPI project template, update the
version of Python to the current patch release. Also update the
version of the ruff pre-commit hook, the GitHub actions versions,
and the install version of Python in GitHub Actions.

Drop the dependencies GitHub Action, which was missed in the
previous change to remove automated dependency updates. We decided
not to use this approach.