Use tox-uv for Python templates
Install the tox-uv plugin to use uv for managing the tox virtual
environments, since it's much faster.

For the FastAPI Safir app, move the dependencies for tox and its
plugins to a new requirements/tox.in file and a corresponding
generated pinnned dependency file, and use the new tox-requirements
support in lsst-sqre/run-tox to use the same file in GitHub Actions.
Let pre-commit float and manage it only in the Makefile. Stop
installing it into tox environments, where it is not needed.
Remove ruff from requirements/dev.in; it's only needed by the lint
target, where it's handled by pre-commit, and it's not useful inside
tox environments.

For SQuaRE PyPI packages, add tox-uv to the Makefile and the GitHub
Actions configuration.