Drop / prefix from doc URLs
Before we had a bug where the OpenAPI docs wouldn't be available because
their URL paths needed a double slash prefix "//" with the default
configuration for path_prefix.

The path_prefix in Config includes a "/" prefix by default, and
app.include_router is set up to use the config's "/" prefix, and so I
think the correct behaviour is for the config to have the path prefix.