Update version handling in Python templates
For both the fastapi_safir_app and square_pypi_package project
templates, add environment variables to the GitHub Actions that hold
the default Python version to use for actions so that it only has to
be updated in one place. Drop the Python version test matrix for
fastapi_safir_app in favor of a single supported Python version.

Stop using neophile for the periodic CI job for fastapi_safir_app
in favor of running make update-deps directly. neophile here was not
adding anything over a simple make invocation.