Add square_pypi_package CI settings
Add timeouts for the test-packaging and pypi jobs following Safir,
and add dependencies for test-packaging since there's likely no
point in running it if the lint, test, or docs builds fail. Allow
the periodic CI job to be run manually. Get the full history when
doing the periodic CI docs check to match the regular docs check.
Make the test-packaging timeout in the periodic CI check match the
one in the regular check.