Update periodic CI notifications
Switch the condition for periodic CI failure notifications from
always() to failure() following the advice in the GitHub
documentation, since always() causes weird problems if it fails.

Add periodic CI notifications to the square_pypi_package template,
similar to those for the fastapi_safir_app template, if the
organization is lsst-sqre.