Add UWS support to FastAPI Safir app template
Add a new configuration option for the FastAPI Safir app project
template that allows the user to indicate whether they are building
a UWS app with a separate backend and a queuing system (defaulting
to no). If that option is enabled, add in the machinery for a UWS
application, including stub models and a stub worker function.

This template uncovered a Ruff diagnostic for unused class method
parameters that would fire on the initial application as generated
from the template. Add an ignore rule for that diagnostic for the
same reason as the similar diagnostic for unused method arguments.