Update the shared Ruff configuration
- Exclude ASYNC109, which required adjusting the comment column
- Drop the S113 exclusion, since its httpx bug was fixed
- Add additional per-file exclusions useful for vertical monorepos
- Drop the McCabe complexity override

The exclusions for vertical monorepos unfortunately appear to require
duplicating the exclusion lists, or at least I don't know a way to
write a glob that matches the paths both with and without a prefix.
Since they have to be duplicated, I made them exact copies so that
future maintenance is an easy cut and paste.

The McCabe complexity threshold increase from 10 to 11 was something
I added for Gafaelfawr, but I found a better way to break apart the
relevant Gafaelfawr function and other packages appear to not have
trouble with a limit of 10.