Add Slack error reporting to fastapi_safir_app
Add support for Slack error reporting of uncaught exceptions to the
`fastapi_safir_app` project template. We are currently adding this
to all of our applications, so there's no reason for it to not be in
the template, even though this may be eventually replaced with Sentry
or some other exception reporting mechanism.