Ignore Ruff's new A005 diagnostic
This diagnoses modules whose names match standard library modules.
We always use relative imports inside applications, and we don't
shadow the standard library module at the top level, so these modules
are not ambiguous and the standard library consumes far too much of
the available namespace to sacrifice all of it.