Mark FastAPI applications as do-not-upload
Add the classifier `Private :: Do Not Upload` to the fastapi_safir_app
template. This tells PyPI to reject the package if it is accidentally
uploaded. We normally do not want to upload service packages to PyPI,
since they are not usable as libraries and often do not attempt to use
a unique namespace. The rare exceptions can manually remove this marker.