Switch fastapi_safir_app template to uv.lock
Adopt `uv.lock` files and uv management of dependencies for the
FastAPI Safir app template. Move all dependencies into `pyproject.toml`
in dependency groups and update the tox configuration accordingly.

Add support for updating the pinned uv version in `Dockerfile` and
in the GitHub CI configuration when running `make update-deps`. The
pinned version of uv will be taken from `uv.lock` (and is guaranteed
to be there because of the dependencies on tox-uv and pre-commit-uv).

Fix the UWS flavor of the template for use of Wobbly, cleaning up a
bunch of problems uncovered by testing a newly-generated template
(renmants of PostgreSQL, different mock API, missing docstring for a
function).

Call `apt-get update` when installing dependency packages as well,
since otherwise there will be no apt lists if installation of base
packages is skipped due to Docker caching.